+++ internal/cli/common/docker/docker.go
@@ -85,3 +85,66 @@ func ComposeCommand() []string {
 	}
 	return []string{"docker-compose"}
 }
+
+// ImageExistsLocally checks if an image exists in the local Docker cache.
+// TODO: Extend Run to support quiet mode so this can use e.Run.
+func (e *Executor) ImageExistsLocally(imageRef string) bool {
+	cmd := exec.Command("docker", "image", "inspect", imageRef)
+	return cmd.Run() == nil
+}
+
+// Pull pulls a Docker image.
+func (e *Executor) Pull(imageRef string) error {
+	if err := e.Run("pull", imageRef); err != nil {
+		return fmt.Errorf("docker pull failed: %w", err)
+	}
+	return nil
+}
+
+// CreateContainer creates a container from an image without starting it and
+// returns the container ID. It tries with an entrypoint override first, then
+// falls back for minimal images without a shell.
+// TODO: Extend Run to support capturing output so this can use e.Run.
+func (e *Executor) CreateContainer(imageRef string) (string, error) {
+	cmd := exec.Command("docker", "create", "--entrypoint", "/bin/sh", imageRef, "-c", "echo")
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		fallback := exec.Command("docker", "create", imageRef)
+		output, err = fallback.CombinedOutput()
+		if err != nil {
+			return "", fmt.Errorf("create container from image: %w: %s", err, strings.TrimSpace(string(output)))
+		}
+	}
+	containerID := strings.TrimSpace(string(output))
+	if containerID == "" {
+		return "", fmt.Errorf("docker create returned empty container id")
+	}
+	return containerID, nil
+}
+
+// CopyFromContainer copies files from a container path to a local path.
+// TODO: Extend Run to support quiet mode so this can use e.Run.
+func (e *Executor) CopyFromContainer(containerID, containerPath, localPath string) error {
+	cmd := exec.Command("docker", "cp", containerID+":"+containerPath, localPath)
+	if e.Verbose {
+		cmd.Stdout = os.Stdout
+		cmd.Stderr = os.Stderr
+		return cmd.Run()
+	}
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("docker cp: %w: %s", err, strings.TrimSpace(string(output)))
+	}
+	return nil
+}
+
+// RemoveContainer removes a container by ID.
+// TODO: Extend Run to support quiet mode so this can use e.Run.
+func (e *Executor) RemoveContainer(containerID string) error {
+	cmd := exec.Command("docker", "rm", containerID)
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		return fmt.Errorf("docker rm: %w: %s", err, strings.TrimSpace(string(output)))
+	}
+	return nil
+}
